Use typographically correct quotes in abstract
It's not feasible to use csquotes in the abstract text, since St. Mary's
requires an abstract to be submitted well in advance of the paper's
completion.  Since this must either be in Word format or *.txt, you can
guess which one I chose.  Having \enquote in there would only confuse
the poor bastard who has to OK my project.